+++ executor/errors.go
@@ -81,7 +81,7 @@ var (
 	ErrLoadParquetFromLocal      = dbterror.ClassExecutor.NewStd(mysql.ErrLoadParquetFromLocal)
 	ErrLoadDataEmptyPath         = dbterror.ClassExecutor.NewStd(mysql.ErrLoadDataEmptyPath)
 	ErrLoadDataUnsupportedFormat = dbterror.ClassExecutor.NewStd(mysql.ErrLoadDataUnsupportedFormat)
-	ErrLoadDataURI               = dbterror.ClassExecutor.NewStd(mysql.ErrLoadDataInvalidURI)
+	ErrLoadDataInvalidURI        = dbterror.ClassExecutor.NewStd(mysql.ErrLoadDataInvalidURI)
 	ErrLoadDataCantAccess        = dbterror.ClassExecutor.NewStd(mysql.ErrLoadDataCantAccess)
 	ErrLoadDataCantRead          = dbterror.ClassExecutor.NewStd(mysql.ErrLoadDataCantRead)
 	ErrLoadDataWrongFormatConfig = dbterror.ClassExecutor.NewStd(mysql.ErrLoadDataWrongFormatConfig)
