+++ executor/utils_test.go
@@ -18,6 +18,7 @@ import (
 	"testing"
 
 	"github.com/pingcap/errors"
+	berrors "github.com/pingcap/tidb/br/pkg/errors"
 	"github.com/stretchr/testify/require"
 )
 
@@ -92,4 +93,13 @@ func TestBatchRetrieverHelper(t *testing.T) {
 	}
 	require.Equal(t, rangeStarts, []int{0})
 	require.Equal(t, rangeEnds, []int{10})
+}
+
+func TestGetMsgFromBRError(t *testing.T) {
+	var berr error = berrors.ErrStorageInvalidConfig
+	require.Equal(t, "[BR:ExternalStorage:ErrStorageInvalidConfig]invalid external storage config", berr.Error())
+	require.Equal(t, "invalid external storage config", getMsgFromBRError(berr))
+	berr = errors.Annotatef(berr, "some message about error reason")
+	require.Equal(t, "some message about error reason: [BR:ExternalStorage:ErrStorageInvalidConfig]invalid external storage config", berr.Error())
+	require.Equal(t, "some message about error reason", getMsgFromBRError(berr))
 }
